Add a dotnet store workflow
This change adds support for `dotnet store` to generate crossgen/R2R
images on the test machine. This is an alternative to AspNet-Install
script, which uses pre-optimized binaries.

See the instructions for details.

I also fixed some issues where we could potentially restore versions of
the shared framework that aren't available. This would block the ability
to do comparisons between builds of the runtime. I fixed this by using the
same 'pinning' technique that's used for the asp.net libraries. See the
readme for details on this.