Workarounds for #33, #34, #35
For #33 and #34 there appears to be something wrong with the authoring of
the AspNetCore.All package. It's pulling in preview1 versions of some
CoreFx packages and running into some bugs with `dotnet store`.
Regardless, we don't use these packages, so I'm working around the problem
by generating the store for the minimal set of packages.

For #35 I don't yet have an answer. Removing the up front install of the
shared framework might mean that the runtime moves slower, but it will
unblock builds for now. Working on getting an answer for this.