Fix crossgen
I made a mistake editing the set of libraries that we crossgen. These
were copied from the csproj, and they use a different MSBuild variable
to set the version oops.

As a result we were getting the 1.0.0 versions in the store and they
didn't match what the app was looking for.